backend/ingestion/requirements.txt

fastapi
uvicorn
python-dotenv
requests
gdown
PyMuPDF
python-docx
openpyxl
python-pptx
Pillow
gradio
moviepy
transformers
torch
sentence-transformers
qdrant-client
faiss-cpu
llama-index
numpy
pandas
